requirements.txt
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.18.0
pandas>=2.0.0
